requirements.txt
requests
praw
discord.py
pyyaml
tinydb
